test(server): split failing test failes
